# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Item Test API",
        "description": "API untuk mengelola data dokter phc",
        "version": "1.0"
    },
    "paths": {
...
        "/api/labtransother": {
            "get": {
                "tags": [
                    "LabTransOther"
                ],
                "summary": "Get all LabTransOthers with optional search",
                "operationId": "3ffe3c986d4d00d359e1c70447d5d930",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Search by exact ID (integer match)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Search by text query (LIKE match), filtering for 'urine' if provided",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LabTransOther"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "LabTransOther"
                ],
                "summary": "Create a new LabTransOther",
                "operationId": "4e1310d511569bf205322f109693e1e4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "LabTransID": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "SupportServiceID": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "SupportServiceNotes": {
                                        "type": "string",
                                        "example": "MRI Scan Required"
                                    },
                                    "CreateBy": {
                                        "type": "string",
                                        "example": "Admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data berhasil disimpan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data berhasil disimpan"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/LabTransOther"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/labtransother/{ID}": {
            "put": {
                "tags": [
                    "LabTransOther"
                ],
                "summary": "Update a specific LabTransOther",
                "operationId": "5184ee5aecdc03fe6ee53b1d191de5b8",
                "parameters": [
                    {
                        "name": "ID",
                        "in": "path",
                        "description": "LabTransOther ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "LabTransID": {
                                        "type": "integer"
                                    },
                                    "SupportServiceID": {
                                        "type": "integer"
                                    },
                                    "SupportServiceNotes": {
                                        "type": "string"
                                    },
                                    "LastModifiedBy": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LabTransOther"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LabTransOther": {
                "required": [
                    "LabTransID",
                    "SupportServiceID"
                ],
                "properties": {
                    "ID": {
                        "type": "integer",
                        "example": 1
                    },
                    "LabTransID": {
                        "type": "integer",
                        "example": 123
                    },
                    "SupportServiceID": {
                        "type": "integer",
                        "example": 456
                    },
                    "SupportServiceNotes": {
                        "type": "string",
                        "example": "Sample notes"
                    },
                    "CreateDate": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "CreateBy": {
                        "type": "string",
                        "example": "admin"
                    },
                    "LastModifiedDate": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "LastModifiedBy": {
                        "type": "string",
                        "example": "admin"
                    },
                    "gcrecord": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
...
        }
    }
}